+++ src/common/dbCommon/mongodbCommon/schemaDB.go
@@ -1,0 +1,10 @@
+package mongodbCommon
+
+import "time"
+
+type RefreshToken struct {
+	Created   time.Time `bson:"created"`
+	Email     string    `bson:"email"`
+	Token     string    `bson:"token"`
+	IsDeleted bool      `bson:"isDeleted"`
+}
