--- src/common/dbCommon/mysql/schemaDB.go
@@ -1,1 +1,0 @@
-package mysql
