+++ src/common/initServer.go
@@ -3,8 +3,8 @@ package common
 import (
 	"fmt"
 	"main/common/awsCommon"
-	"main/common/dbCommon/mongodb"
-	"main/common/dbCommon/mysql"
+	"main/common/dbCommon/mongodbCommon"
+	"main/common/dbCommon/mysqlCommon"
 	"main/common/envCommon"
 	"main/common/oauthCommon/google"
 )
@@ -23,12 +23,12 @@ func InitServer() error {
 		return err
 	}
 	//TODO 몽고디비 대책이 필요 계속 뻑나네..
-	if err := mongodb.InitMongoDB(); err != nil {
+	if err := mongodbCommon.InitMongoDB(); err != nil {
 		fmt.Sprintf("mongoDB 초기화 에러 : %s", err.Error())
 		return err
 	}
-	if err := mysql.InitMySQL(); err != nil {
-		fmt.Sprintf("mysql 초기화 에러 : %s", err.Error())
+	if err := mysqlCommon.InitMySQL(); err != nil {
+		fmt.Sprintf("mysqlCommon 초기화 에러 : %s", err.Error())
 		return err
 	}
 	return nil
