+++ src/features/oauth/google/handler/callbackGoogleOAuthHandler.go
@@ -6,7 +6,7 @@ import (
 	"github.com/labstack/echo/v4"
 	"golang.org/x/oauth2"
 	"io/ioutil"
-	"main/common/dbCommon/mongodb"
+	"main/common/dbCommon/mongodbCommon"
 	"main/common/oauthCommon/google"
 	"main/features/oauth/google/model/response"
 	"main/features/oauth/google/repository"
@@ -21,7 +21,7 @@ type CallbackGoogleOAuthHandler struct {
 }
 
 func NewCallbackGoogleOAuthHandler() *CallbackGoogleOAuthHandler {
-	return &CallbackGoogleOAuthHandler{UseCase: usecase.NewCallbackGoogleOAuthUseCase(repository.NewCallbackGoogleOAuthRepository(mongodb.TokenCollection))}
+	return &CallbackGoogleOAuthHandler{UseCase: usecase.NewCallbackGoogleOAuthUseCase(repository.NewCallbackGoogleOAuthRepository(mongodbCommon.TokenCollection))}
 }
 
 // google signin callback
