+++ src/features/oauth/google/repository/callbackGoogleOAuthRepository.go
@@ -5,7 +5,8 @@ import (
 	"fmt"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo"
-	"main/common/dbCommon/mongodb"
+	"main/common/dbCommon/mongodbCommon"
+	"main/common/dbCommon/mysqlCommon"
 	"main/common/oauthCommon/google"
 	_interface "main/features/oauth/google/usecase/interface"
 )
@@ -19,7 +20,7 @@ func (cc *CallbackGoogleOAuthRepository) CallbackGoogle() error {
 	return nil
 }
 
-func (cc *CallbackGoogleOAuthRepository) CreateRefreshToken(token mongodb.RefreshToken) error {
+func (cc *CallbackGoogleOAuthRepository) CreateRefreshToken(token mongodbCommon.RefreshToken) error {
 	ctx := context.TODO()
 	_, err := cc.TokenCollection.InsertOne(ctx, token)
 	if err != nil {
@@ -36,5 +37,23 @@ func (cc *CallbackGoogleOAuthRepository) DeleteAllRefreshToken(authUser google.U
 		return err
 	}
 	fmt.Println(result.DeletedCount)
+	return nil
+}
+
+func (cc *CallbackGoogleOAuthRepository) FindOneUser(authUser google.User) (bool, error) {
+	var email string
+	err := mysqlCommon.MysqlDB.QueryRow("SELECT email FROM user WHERE email = ?", authUser.Email).Scan(&email)
+	if err != nil {
+		if err.Error() != "sql: no rows in result set" {
+			fmt.Println(err)
+			return false, err
+		}
+		return false, nil
+	}
+	return true, nil
+}
+
+func (cc *CallbackGoogleOAuthRepository) CreateUser(authUser google.User) error {
+
 	return nil
 }
